Adding time/frame conversion utils and tests

+++ src/duke/playback/Player.h
@@ -12,17 +12,17 @@ struct Player: public NonCopyable {
 		m_Timeline = timeline;
 		cue(m_Timeline.empty() ? 0 : m_Timeline.getRange().first);
 	}
-	void play(size_t frame, int rate = 1);
+	void play(uint32_t frame, int rate = 1);
 	void stop();
-	inline void cue(size_t frame) {
+	inline void cue(uint32_t frame) {
 		play(frame, 0);
 	}
-	inline size_t getCurrentFrame() const {
+	inline uint32_t getCurrentFrame() const {
 		return m_CurrentFrame;
 	}
 private:
 	Timeline m_Timeline;
-	size_t m_CurrentFrame;
+	uint32_t m_CurrentFrame;
 	uint64_t m_ElapsedMicroseconds;
 };
